feat: Add unit tests and add missing features

+++ db/sqlc/models.go
@@ -9,7 +9,7 @@ import (
 )
 
 type Account struct {
-	ID        int32     `json:"id"`
+	ID        int64     `json:"id"`
 	Owner     string    `json:"owner"`
 	Balance   int64     `json:"balance"`
 	Currency  string    `json:"currency"`
@@ -17,17 +17,17 @@ type Account struct {
 }
 
 type Entry struct {
-	ID        int32 `json:"id"`
-	AccountID int32 `json:"account_id"`
+	ID        int64 `json:"id"`
+	AccountID int64 `json:"account_id"`
 	// can be -ve or +ve
 	Amount    int64     `json:"amount"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
 type Transfer struct {
-	ID            int32 `json:"id"`
-	FromAccountID int32 `json:"from_account_id"`
-	ToAccountID   int32 `json:"to_account_id"`
+	ID            int64 `json:"id"`
+	FromAccountID int64 `json:"from_account_id"`
+	ToAccountID   int64 `json:"to_account_id"`
 	// must be +ve
 	Amount    int64     `json:"amount"`
 	CreatedAt time.Time `json:"created_at"`
